requirements.txt
flask
flask-cors
google-generativeai
deep-translator
gunicorn  